It's faster again; I'm pretty hype
Apparently, a big bottleneck in speed came from converting the characters to numbers. Previously,
it would search through the entire charts file until it found the right character. This happened
for every character.

Now, it keeps track of all of the already found numbers that correspond to characters in an array.
Each character's corresponding number can be found at that character's ASCII value in the array.
For example, once the number for "A" has been found, it's stored at index 65 in the array, and
creboot will get it from there until the encrypt/decrypt function is re-run.